backend/requirements.txt
flask
mysql-connector-python
apscheduler
akshare
requests
pandas
flask-cors
